backend/requirements.txt
fastapi[standard]
uvicorn
bcrypt
python-jose[cryptography]
python-multipart
